requirements.txt
scrapy>=2.8.0
google-api-python-client>=2.70.0
openai>=0.27.0
python-dotenv>=1.0.0